Module embassy_stm32::dac

Digital to Analog Converter (DAC)

Structs

Dac

DAC driver.

DacChannel

Driver for a single DAC channel.

Enums

DualValue

Dual 8 or 12 bit values that can be output by the DAC channels 1 and 2 simultaneously.

Mode

Operating mode for DAC channel

TriggerSel

Trigger selection for L4+, L5, U5, H7.

Value

Single 8 or 12 bit value that can be output by the DAC.

ValueArray

Array variant of Value.

Traits

DacDma1

DacDma1 DMA request trait

DacDma2

DacDma2 DMA request trait

DacPin

Marks a pin that can be used with the DAC

Instance

DAC instance.

Type Aliases

DacCh1

DAC channel 1 type alias.

DacCh2

DAC channel 2 type alias.