Module embassy_stm32::dcmi

Digital Camera Interface (DCMI)

Structs

Config

DCMI configuration.

Dcmi

DCMI driver.

InterruptHandler

Interrupt handler.

Enums

Error

DCMI error.

HSyncDataInvalidLevel

The level on the VSync pin when the data is not valid on the parallel interface.

PixelClockPolarity

VSyncDataInvalidLevel

The level on the VSync pin when the data is not valid on the parallel interface.

Traits

D0Pin

D0Pin pin trait

D1Pin

D1Pin pin trait

D2Pin

D2Pin pin trait

D3Pin

D3Pin pin trait

D4Pin

D4Pin pin trait

D5Pin

D5Pin pin trait

D6Pin

D6Pin pin trait

D7Pin

D7Pin pin trait

D8Pin

D8Pin pin trait

D9Pin

D9Pin pin trait

D10Pin

D10Pin pin trait

D11Pin

D11Pin pin trait

D12Pin

D12Pin pin trait

D13Pin

D13Pin pin trait

FrameDma

FrameDma DMA request trait

HSyncPin

HSyncPin pin trait

Instance

DCMI instance.

PixClkPin

PixClkPin pin trait

VSyncPin

VSyncPin pin trait