Struct embassy_stm32::dma::WritableRingBuffer

pub struct WritableRingBuffer<'a, C: Channel, W: Word> { /* private fields */ }

Ringbuffer for writing data using DMA circular mode.

Implementations

impl<'a, C: Channel, W: Word> WritableRingBuffer<'a, C, W>

pub unsafe fn new( channel: impl Peripheral<P = C> + 'a, _request: Request, peri_addr: *mut W, buffer: &'a mut [W], options: TransferOptions ) -> Self

Create a new ring buffer.

pub fn start(&mut self)

Start the ring buffer operation.

You must call this after creating it for it to work.

pub fn clear(&mut self)

Clear all data in the ring buffer.

pub fn write(&mut self, buf: &[W]) -> Result<(usize, usize), OverrunError>

Write elements from the ring buffer Return a tuple of the length written and the length remaining in the buffer

pub async fn write_exact(&mut self, buffer: &[W]) -> Result<usize, OverrunError>

Write an exact number of elements to the ringbuffer.

pub const fn capacity(&self) -> usize

The capacity of the ringbuffer

pub fn set_waker(&mut self, waker: &Waker)

Set a waker to be woken when at least one byte is received.

pub fn request_stop(&mut self)

Request DMA to stop.

This doesn’t immediately stop the transfer, you have to wait until is_running returns false.

pub fn is_running(&mut self) -> bool

Return whether DMA is still running.

If this returns false, it can be because either the transfer finished, or it was requested to stop early with request_stop.