Struct embassy_stm32::flash::Bank2Region

pub struct Bank2Region<'d, MODE = Async>(pub &'static FlashRegion, _, _);

Tuple Fields
0: &'static FlashRegion
Implementations

impl<MODE> Bank2Region<'_, MODE>

pub fn blocking_read( &mut self, offset: u32, bytes: &mut [u8] ) -> Result<(), Error>

Blocking read.

NOTE: offset is an offset from the flash start, NOT an absolute address. For example, to read address 0x0800_1234 you have to use offset 0x1234.

impl Bank2Region<'_, Blocking>

pub fn blocking_write(&mut self, offset: u32, bytes: &[u8]) -> Result<(), Error>

Blocking write.

NOTE: offset is an offset from the flash start, NOT an absolute address. For example, to write address 0x0800_1234 you have to use offset 0x1234.

pub fn blocking_erase(&mut self, from: u32, to: u32) -> Result<(), Error>

Blocking erase.

NOTE: from and to are offsets from the flash start, NOT an absolute address. For example, to erase address 0x0801_0000 you have to use offset 0x1_0000.