Struct embassy_stm32::gpio::Output

pub struct Output<'d, T: Pin> { /* private fields */ }

GPIO output driver.

Note that pins will return to their floating state when Output is dropped. If pins should retain their state indefinitely, either keep ownership of the Output, or pass it to core::mem::forget.

Implementations

impl<'d, T: Pin> Output<'d, T>

pub fn new( pin: impl Peripheral<P = T> + 'd, initial_output: Level, speed: Speed ) -> Self

Create GPIO output driver for a Pin with the provided Level and Speed configuration.

pub fn degrade(self) -> Output<'d, AnyPin>

Type-erase (degrade) this pin into an AnyPin.

This converts pin singletons (PA5, PB6, …), which are all different types, into the same type. It is useful for creating arrays of pins, or avoiding generics.

pub fn set_high(&mut self)

Set the output as high.

pub fn set_low(&mut self)

Set the output as low.

pub fn set_level(&mut self, level: Level)

Set the output level.

pub fn is_set_high(&self) -> bool

Is the output pin set as high?

pub fn is_set_low(&self) -> bool

Is the output pin set as low?

pub fn get_output_level(&self) -> Level

What level output is set to

pub fn toggle(&mut self)

Toggle pin output