Trait embassy_stm32::gpio::Pin

pub trait Pin: Peripheral<P = Self> + Into<AnyPin> + Pin + Sized + 'static {
    type ExtiChannel: Channel;

    // Provided methods
    fn pin(&self) -> u8 { ... }
    fn port(&self) -> u8 { ... }
    fn degrade(self) -> AnyPin { ... }
}

GPIO pin trait.

Required Associated Types

type ExtiChannel: Channel

EXTI channel assigned to this pin.

For example, PC4 uses EXTI4.

Provided Methods

fn pin(&self) -> u8

Number of the pin within the port (0..31)

fn port(&self) -> u8

Port of the pin

fn degrade(self) -> AnyPin

Type-erase (degrade) this pin into an AnyPin.

This converts pin singletons (PA5, PB6, …), which are all different types, into the same type. It is useful for creating arrays of pins, or avoiding generics.

Object Safety

This trait is not object safe.