Struct embassy_stm32::i2c::I2c

pub struct I2c<'d, T: Instance, TXDMA = NoDma, RXDMA = NoDma> { /* private fields */ }

I2C driver.

Implementations

impl<'d, T: Instance, TXDMA, RXDMA> I2c<'d, T, TXDMA, RXDMA>

pub async fn write(&mut self, address: u8, write: &[u8]) -> Result<(), Error>
where TXDMA: TxDma<T>,

Write.

pub async fn write_vectored( &mut self, address: u8, write: &[&[u8]] ) -> Result<(), Error>
where TXDMA: TxDma<T>,

Write multiple buffers.

The buffers are concatenated in a single write transaction.

pub async fn read( &mut self, address: u8, buffer: &mut [u8] ) -> Result<(), Error>
where RXDMA: RxDma<T>,

Read.

pub async fn write_read( &mut self, address: u8, write: &[u8], read: &mut [u8] ) -> Result<(), Error>
where TXDMA: TxDma<T>, RXDMA: RxDma<T>,

Write, restart, read.

pub fn blocking_read( &mut self, address: u8, read: &mut [u8] ) -> Result<(), Error>

Blocking read.

pub fn blocking_write(&mut self, address: u8, write: &[u8]) -> Result<(), Error>

Blocking write.

pub fn blocking_write_read( &mut self, address: u8, write: &[u8], read: &mut [u8] ) -> Result<(), Error>

Blocking write, restart, read.

pub fn blocking_write_vectored( &mut self, address: u8, write: &[&[u8]] ) -> Result<(), Error>

Blocking write multiple buffers.

The buffers are concatenated in a single write transaction.

impl<'d, T: Instance, TXDMA, RXDMA> I2c<'d, T, TXDMA, RXDMA>

pub fn new( peri: impl Peripheral<P = T> + 'd, scl: impl Peripheral<P = impl SclPin<T>> + 'd, sda: impl Peripheral<P = impl SdaPin<T>> + 'd, _irq: impl Binding<T::EventInterrupt, EventInterruptHandler<T>> + Binding<T::ErrorInterrupt, ErrorInterruptHandler<T>> + 'd, tx_dma: impl Peripheral<P = TXDMA> + 'd, rx_dma: impl Peripheral<P = RXDMA> + 'd, freq: Hertz, config: Config ) -> Self

Create a new I2C driver.