Module embassy_stm32::interrupt::typelevel

Type-level interrupt infrastructure.

This module contains one type per interrupt. This is used for checking at compile time that the interrupts are correctly bound to HAL drivers.

As an end user, you shouldn’t need to use this module directly. Use the crate::bind_interrupts! macro to bind interrupts, and the [crate::interrupt] module to manually register interrupt handlers and manipulate interrupts directly (pending/unpending, enabling/disabling, setting the priority, etc…)

Enums

ADC

ADC typelevel interrupt.

BDMA1

BDMA1 typelevel interrupt.

BDMA2_CHANNEL0

BDMA2_CHANNEL0 typelevel interrupt.

BDMA2_CHANNEL1

BDMA2_CHANNEL1 typelevel interrupt.

BDMA2_CHANNEL2

BDMA2_CHANNEL2 typelevel interrupt.

BDMA2_CHANNEL3

BDMA2_CHANNEL3 typelevel interrupt.

BDMA2_CHANNEL4

BDMA2_CHANNEL4 typelevel interrupt.

BDMA2_CHANNEL5

BDMA2_CHANNEL5 typelevel interrupt.

BDMA2_CHANNEL6

BDMA2_CHANNEL6 typelevel interrupt.

BDMA2_CHANNEL7

BDMA2_CHANNEL7 typelevel interrupt.

CEC

CEC typelevel interrupt.

CRS

CRS typelevel interrupt.

DAC2

DAC2 typelevel interrupt.

DCMI_PSSI

DCMI_PSSI typelevel interrupt.

DFSDM1_FLT0

DFSDM1_FLT0 typelevel interrupt.

DFSDM1_FLT1

DFSDM1_FLT1 typelevel interrupt.

DFSDM1_FLT2

DFSDM1_FLT2 typelevel interrupt.

DFSDM1_FLT3

DFSDM1_FLT3 typelevel interrupt.

DFSDM1_FLT4

DFSDM1_FLT4 typelevel interrupt.

DFSDM1_FLT5

DFSDM1_FLT5 typelevel interrupt.

DFSDM1_FLT6

DFSDM1_FLT6 typelevel interrupt.

DFSDM1_FLT7

DFSDM1_FLT7 typelevel interrupt.

DFSDM2

DFSDM2 typelevel interrupt.

DMA1_STREAM0

DMA1_STREAM0 typelevel interrupt.

DMA1_STREAM1

DMA1_STREAM1 typelevel interrupt.

DMA1_STREAM2

DMA1_STREAM2 typelevel interrupt.

DMA1_STREAM3

DMA1_STREAM3 typelevel interrupt.

DMA1_STREAM4

DMA1_STREAM4 typelevel interrupt.

DMA1_STREAM5

DMA1_STREAM5 typelevel interrupt.

DMA1_STREAM6

DMA1_STREAM6 typelevel interrupt.

DMA1_STREAM7

DMA1_STREAM7 typelevel interrupt.

DMA2D

DMA2D typelevel interrupt.

DMA2_STREAM0

DMA2_STREAM0 typelevel interrupt.

DMA2_STREAM1

DMA2_STREAM1 typelevel interrupt.

DMA2_STREAM2

DMA2_STREAM2 typelevel interrupt.

DMA2_STREAM3

DMA2_STREAM3 typelevel interrupt.

DMA2_STREAM4

DMA2_STREAM4 typelevel interrupt.

DMA2_STREAM5

DMA2_STREAM5 typelevel interrupt.

DMA2_STREAM6

DMA2_STREAM6 typelevel interrupt.

DMA2_STREAM7

DMA2_STREAM7 typelevel interrupt.

DMAMUX1_OVR

DMAMUX1_OVR typelevel interrupt.

DMAMUX2_OVR

DMAMUX2_OVR typelevel interrupt.

DTS

DTS typelevel interrupt.

ECC

ECC typelevel interrupt.

EXTI0

EXTI0 typelevel interrupt.

EXTI1

EXTI1 typelevel interrupt.

EXTI2

EXTI2 typelevel interrupt.

EXTI3

EXTI3 typelevel interrupt.

EXTI4

EXTI4 typelevel interrupt.

EXTI9_5

EXTI9_5 typelevel interrupt.

EXTI15_10

EXTI15_10 typelevel interrupt.

FDCAN1_IT0

FDCAN1_IT0 typelevel interrupt.

FDCAN1_IT1

FDCAN1_IT1 typelevel interrupt.

FDCAN2_IT0

FDCAN2_IT0 typelevel interrupt.

FDCAN2_IT1

FDCAN2_IT1 typelevel interrupt.

FDCAN_CAL

FDCAN_CAL typelevel interrupt.

FLASH

FLASH typelevel interrupt.

FMC

FMC typelevel interrupt.

FPU

FPU typelevel interrupt.

GFXMMU

GFXMMU typelevel interrupt.

HSEM1

HSEM1 typelevel interrupt.

I2C1_ER

I2C1_ER typelevel interrupt.

I2C1_EV

I2C1_EV typelevel interrupt.

I2C2_ER

I2C2_ER typelevel interrupt.

I2C2_EV

I2C2_EV typelevel interrupt.

I2C3_ER

I2C3_ER typelevel interrupt.

I2C3_EV

I2C3_EV typelevel interrupt.

I2C4_ER

I2C4_ER typelevel interrupt.

I2C4_EV

I2C4_EV typelevel interrupt.

JPEG

JPEG typelevel interrupt.

LPTIM1

LPTIM1 typelevel interrupt.

LPTIM2

LPTIM2 typelevel interrupt.

LPTIM3

LPTIM3 typelevel interrupt.

LPUART1

LPUART1 typelevel interrupt.

LTDC

LTDC typelevel interrupt.

LTDC_ER

LTDC_ER typelevel interrupt.

MDIOS

MDIOS typelevel interrupt.

MDIOS_WKUP

MDIOS_WKUP typelevel interrupt.

MDMA

MDMA typelevel interrupt.

OCTOSPI1

OCTOSPI1 typelevel interrupt.

OCTOSPI2

OCTOSPI2 typelevel interrupt.

OTG_HS

OTG_HS typelevel interrupt.

OTG_HS_EP1_IN

OTG_HS_EP1_IN typelevel interrupt.

OTG_HS_EP1_OUT

OTG_HS_EP1_OUT typelevel interrupt.

OTG_HS_WKUP

OTG_HS_WKUP typelevel interrupt.

PVD_PVM

PVD_PVM typelevel interrupt.

RCC

RCC typelevel interrupt.

RNG

RNG typelevel interrupt.

RTC_ALARM

RTC_ALARM typelevel interrupt.

RTC_TAMP_STAMP_CSS_LSE

RTC_TAMP_STAMP_CSS_LSE typelevel interrupt.

RTC_WKUP

RTC_WKUP typelevel interrupt.

SAI1

SAI1 typelevel interrupt.

SAI2

SAI2 typelevel interrupt.

SDMMC1

SDMMC1 typelevel interrupt.

SDMMC2

SDMMC2 typelevel interrupt.

SPDIF_RX

SPDIF_RX typelevel interrupt.

SPI1

SPI1 typelevel interrupt.

SPI2

SPI2 typelevel interrupt.

SPI3

SPI3 typelevel interrupt.

SPI4

SPI4 typelevel interrupt.

SPI5

SPI5 typelevel interrupt.

SPI6

SPI6 typelevel interrupt.

SWPMI1

SWPMI1 typelevel interrupt.

TIM1_BRK

TIM1_BRK typelevel interrupt.

TIM1_CC

TIM1_CC typelevel interrupt.

TIM1_TRG_COM

TIM1_TRG_COM typelevel interrupt.

TIM1_UP

TIM1_UP typelevel interrupt.

TIM2

TIM2 typelevel interrupt.

TIM3

TIM3 typelevel interrupt.

TIM4

TIM4 typelevel interrupt.

TIM5

TIM5 typelevel interrupt.

TIM6_DAC

TIM6_DAC typelevel interrupt.

TIM7

TIM7 typelevel interrupt.

TIM8_BRK_TIM12

TIM8_BRK_TIM12 typelevel interrupt.

TIM8_CC

TIM8_CC typelevel interrupt.

TIM8_TRG_COM_TIM14

TIM8_TRG_COM_TIM14 typelevel interrupt.

TIM8_UP_TIM13

TIM8_UP_TIM13 typelevel interrupt.

TIM15

TIM15 typelevel interrupt.

TIM16

TIM16 typelevel interrupt.

TIM17

TIM17 typelevel interrupt.

UART4

UART4 typelevel interrupt.

UART5

UART5 typelevel interrupt.

UART7

UART7 typelevel interrupt.

UART8

UART8 typelevel interrupt.

UART9

UART9 typelevel interrupt.

USART1

USART1 typelevel interrupt.

USART2

USART2 typelevel interrupt.

USART3

USART3 typelevel interrupt.

USART6

USART6 typelevel interrupt.

USART10

USART10 typelevel interrupt.

WAKEUP_PIN

WAKEUP_PIN typelevel interrupt.

WWDG

WWDG typelevel interrupt.

WWDG_RST

WWDG_RST typelevel interrupt.

Traits

Binding

Compile-time assertion that an interrupt has been bound to a handler.

Handler

Interrupt handler trait.

Interrupt

Type-level interrupt.