Trait embassy_stm32::interrupt::typelevel::Interrupt

pub trait Interrupt: Interrupt {
    const IRQ: Interrupt;

    // Provided methods
    unsafe fn enable() { ... }
    fn disable() { ... }
    fn is_enabled() -> bool { ... }
    fn is_pending() -> bool { ... }
    fn pend() { ... }
    fn unpend() { ... }
    fn get_priority() -> Priority { ... }
    fn set_priority(prio: Priority) { ... }
    fn set_priority_with_cs(cs: CriticalSection<'_>, prio: Priority) { ... }
}

Type-level interrupt.

This trait is implemented for all typelevel interrupt types in this module.

Required Associated Constants

const IRQ: Interrupt

Interrupt enum variant.

This allows going from typelevel interrupts (one type per interrupt) to non-typelevel interrupts (a single Interrupt enum type, with one variant per interrupt).

Provided Methods

unsafe fn enable()

Enable the interrupt.

fn disable()

Disable the interrupt.

fn is_enabled() -> bool

Check if interrupt is enabled.

fn is_pending() -> bool

Check if interrupt is pending.

fn pend()

Set interrupt pending.

fn unpend()

Unset interrupt pending.

fn get_priority() -> Priority

Get the priority of the interrupt.

fn set_priority(prio: Priority)

Set the interrupt priority.

fn set_priority_with_cs(cs: CriticalSection<'_>, prio: Priority)

Set the interrupt priority with an already-acquired critical section

Object Safety

This trait is not object safe.