Struct embassy_stm32::peripherals::DCMI

pub struct DCMI { /* private fields */ }

DCMI peripheral

Implementations

impl DCMI

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.