Struct embassy_stm32::peripherals::DMA2_CH1

pub struct DMA2_CH1 { /* private fields */ }

DMA2_CH1 peripheral

Implementations

impl DMA2_CH1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.