Struct embassy_stm32::peripherals::EXTI1

pub struct EXTI1 { /* private fields */ }

EXTI1 peripheral

Implementations

impl EXTI1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.