Struct embassy_stm32::peripherals::FMC

pub struct FMC { /* private fields */ }

FMC peripheral

Implementations

impl FMC

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.