Struct embassy_stm32::peripherals::PK6

pub struct PK6 { /* private fields */ }

PK6 peripheral

Implementations

impl PK6

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.