Struct embassy_stm32::peripherals::SDMMC1

pub struct SDMMC1 { /* private fields */ }

SDMMC1 peripheral

Implementations

impl SDMMC1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.