Struct embassy_stm32::peripherals::TIM1

pub struct TIM1 { /* private fields */ }

TIM1 peripheral

Implementations

impl TIM1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.