Struct embassy_stm32::peripherals::USB_OTG_HS

pub struct USB_OTG_HS { /* private fields */ }

USB_OTG_HS peripheral

Implementations

impl USB_OTG_HS

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.