Enum embassy_stm32::rcc::SupplyConfig

pub enum SupplyConfig {
    Default,
    LDO,
    DirectSMPS,
    SMPSLDO(SMPSSupplyVoltage),
    SMPSExternalLDO(SMPSSupplyVoltage),
    SMPSExternalLDOBypass(SMPSSupplyVoltage),
    SMPSDisabledLDOBypass,
}

Power supply configuration See RM0433 Rev 4 7.4

Variants

Default

Default power supply configuration. V CORE Power Domains are supplied from the LDO according to VOS. SMPS step-down converter enabled at 1.2V, may be used to supply the LDO.

LDO

Power supply configuration using the LDO. V CORE Power Domains are supplied from the LDO according to VOS. LDO power mode (Main, LP, Off) will follow system low-power modes. SMPS step-down converter disabled.

DirectSMPS

Power supply configuration directly from the SMPS step-down converter. V CORE Power Domains are supplied from SMPS step-down converter according to VOS. LDO bypassed. SMPS step-down converter power mode (MR, LP, Off) will follow system low-power modes.

SMPSLDO(SMPSSupplyVoltage)

Power supply configuration from the SMPS step-down converter, that supplies the LDO. V CORE Power Domains are supplied from the LDO according to VOS LDO power mode (Main, LP, Off) will follow system low-power modes. SMPS step-down converter enabled according to SDLEVEL, and supplies the LDO. SMPS step-down converter power mode (MR, LP, Off) will follow system low-power modes.

SMPSExternalLDO(SMPSSupplyVoltage)

Power supply configuration from SMPS supplying external circuits and potentially the LDO. V CORE Power Domains are supplied from voltage regulator according to VOS LDO power mode (Main, LP, Off) will follow system low-power modes. SMPS step-down converter enabled according to SDLEVEL used to supply external circuits and may supply the LDO. SMPS step-down converter forced ON in MR mode.

SMPSExternalLDOBypass(SMPSSupplyVoltage)

Power supply configuration from SMPS supplying external circuits and bypassing the LDO. V CORE supplied from external source SMPS step-down converter enabled according to SDLEVEL used to supply external circuits and may supply the external source for V CORE . SMPS step-down converter forced ON in MR mode.

SMPSDisabledLDOBypass

Power supply configuration from an external source, SMPS disabled and the LDO bypassed. V CORE supplied from external source SMPS step-down converter disabled and LDO bypassed, voltage monitoring still active.