Struct embassy_stm32::spi::Spi

pub struct Spi<'d, T: Instance, Tx, Rx> { /* private fields */ }

SPI driver.

Implementations

impl<'d, T: Instance, Tx, Rx> Spi<'d, T, Tx, Rx>

pub fn new( peri: impl Peripheral<P = T> + 'd, sck: impl Peripheral<P = impl SckPin<T>> + 'd, mosi: impl Peripheral<P = impl MosiPin<T>> + 'd, miso: impl Peripheral<P = impl MisoPin<T>> + 'd, txdma: impl Peripheral<P = Tx> + 'd, rxdma: impl Peripheral<P = Rx> + 'd, config: Config ) -> Self

Create a new SPI driver.

pub fn new_rxonly( peri: impl Peripheral<P = T> + 'd, sck: impl Peripheral<P = impl SckPin<T>> + 'd, miso: impl Peripheral<P = impl MisoPin<T>> + 'd, txdma: impl Peripheral<P = Tx> + 'd, rxdma: impl Peripheral<P = Rx> + 'd, config: Config ) -> Self

Create a new SPI driver, in RX-only mode (only MISO pin, no MOSI).

pub fn new_txonly( peri: impl Peripheral<P = T> + 'd, sck: impl Peripheral<P = impl SckPin<T>> + 'd, mosi: impl Peripheral<P = impl MosiPin<T>> + 'd, txdma: impl Peripheral<P = Tx> + 'd, rxdma: impl Peripheral<P = Rx> + 'd, config: Config ) -> Self

Create a new SPI driver, in TX-only mode (only MOSI pin, no MISO).

pub fn new_txonly_nosck( peri: impl Peripheral<P = T> + 'd, mosi: impl Peripheral<P = impl MosiPin<T>> + 'd, txdma: impl Peripheral<P = Tx> + 'd, rxdma: impl Peripheral<P = Rx> + 'd, config: Config ) -> Self

Create a new SPI driver, in TX-only mode, without SCK pin.

This can be useful for bit-banging non-SPI protocols.

pub fn set_config(&mut self, config: &Config) -> Result<(), ()>

Reconfigures it with the supplied config.

pub fn get_current_config(&self) -> Config

Get current SPI configuration.

pub async fn write<W: Word>(&mut self, data: &[W]) -> Result<(), Error>
where Tx: TxDma<T>,

SPI write, using DMA.

pub async fn read<W: Word>(&mut self, data: &mut [W]) -> Result<(), Error>
where Tx: TxDma<T>, Rx: RxDma<T>,

SPI read, using DMA.

pub async fn transfer<W: Word>( &mut self, read: &mut [W], write: &[W] ) -> Result<(), Error>
where Tx: TxDma<T>, Rx: RxDma<T>,

Bidirectional transfer, using DMA.

This transfers both buffers at the same time, so it is NOT equivalent to write followed by read.

The transfer runs for max(read.len(), write.len()) bytes. If read is shorter extra bytes are ignored. If write is shorter it is padded with zero bytes.

pub async fn transfer_in_place<W: Word>( &mut self, data: &mut [W] ) -> Result<(), Error>
where Tx: TxDma<T>, Rx: RxDma<T>,

In-place bidirectional transfer, using DMA.

This writes the contents of data on MOSI, and puts the received data on MISO in data, at the same time.

pub fn blocking_write<W: Word>(&mut self, words: &[W]) -> Result<(), Error>

Blocking write.

pub fn blocking_read<W: Word>(&mut self, words: &mut [W]) -> Result<(), Error>

Blocking read.

pub fn blocking_transfer_in_place<W: Word>( &mut self, words: &mut [W] ) -> Result<(), Error>

Blocking in-place bidirectional transfer.

This writes the contents of data on MOSI, and puts the received data on MISO in data, at the same time.

pub fn blocking_transfer<W: Word>( &mut self, read: &mut [W], write: &[W] ) -> Result<(), Error>

Blocking bidirectional transfer.

This transfers both buffers at the same time, so it is NOT equivalent to write followed by read.

The transfer runs for max(read.len(), write.len()) bytes. If read is shorter extra bytes are ignored. If write is shorter it is padded with zero bytes.