Struct embassy_stm32::Peripherals

pub struct Peripherals {

    pub ADC1: ADC1,
    pub ADC2: ADC2,
    pub ADC_COMMON: ADC_COMMON,
    pub BDMA1: BDMA1,
    pub BDMA2: BDMA2,
    pub CRC: CRC,
    pub CRS: CRS,
    pub DAC1: DAC1,
    pub DAC2: DAC2,
    pub DBGMCU: DBGMCU,
    pub DCMI: DCMI,
    pub DMA1: DMA1,
    pub DMA2: DMA2,
    pub DMA2D: DMA2D,
    pub DMAMUX1: DMAMUX1,
    pub DMAMUX2: DMAMUX2,
    pub FDCAN1: FDCAN1,
    pub FDCAN2: FDCAN2,
    pub FDCANRAM: FDCANRAM,
    pub FLASH: FLASH,
    pub FMC: FMC,
    pub PA0: PA0,
    pub PA1: PA1,
    pub PA2: PA2,
    pub PA3: PA3,
    pub PA4: PA4,
    pub PA5: PA5,
    pub PA6: PA6,
    pub PA7: PA7,
    pub PA8: PA8,
    pub PA9: PA9,
    pub PA10: PA10,
    pub PA11: PA11,
    pub PA12: PA12,
    pub PA13: PA13,
    pub PA14: PA14,
    pub PA15: PA15,
    pub PB0: PB0,
    pub PB1: PB1,
    pub PB2: PB2,
    pub PB3: PB3,
    pub PB4: PB4,
    pub PB5: PB5,
    pub PB6: PB6,
    pub PB7: PB7,
    pub PB8: PB8,
    pub PB9: PB9,
    pub PB10: PB10,
    pub PB11: PB11,
    pub PB12: PB12,
    pub PB13: PB13,
    pub PB14: PB14,
    pub PB15: PB15,
    pub PC0: PC0,
    pub PC1: PC1,
    pub PC2: PC2,
    pub PC3: PC3,
    pub PC4: PC4,
    pub PC5: PC5,
    pub PC6: PC6,
    pub PC7: PC7,
    pub PC8: PC8,
    pub PC9: PC9,
    pub PC10: PC10,
    pub PC11: PC11,
    pub PC12: PC12,
    pub PC13: PC13,
    pub PC14: PC14,
    pub PC15: PC15,
    pub PD0: PD0,
    pub PD1: PD1,
    pub PD2: PD2,
    pub PD3: PD3,
    pub PD4: PD4,
    pub PD5: PD5,
    pub PD6: PD6,
    pub PD7: PD7,
    pub PD8: PD8,
    pub PD9: PD9,
    pub PD10: PD10,
    pub PD11: PD11,
    pub PD12: PD12,
    pub PD13: PD13,
    pub PD14: PD14,
    pub PD15: PD15,
    pub PE0: PE0,
    pub PE1: PE1,
    pub PE2: PE2,
    pub PE3: PE3,
    pub PE4: PE4,
    pub PE5: PE5,
    pub PE6: PE6,
    pub PE7: PE7,
    pub PE8: PE8,
    pub PE9: PE9,
    pub PE10: PE10,
    pub PE11: PE11,
    pub PE12: PE12,
    pub PE13: PE13,
    pub PE14: PE14,
    pub PE15: PE15,
    pub PF0: PF0,
    pub PF1: PF1,
    pub PF2: PF2,
    pub PF3: PF3,
    pub PF4: PF4,
    pub PF5: PF5,
    pub PF6: PF6,
    pub PF7: PF7,
    pub PF8: PF8,
    pub PF9: PF9,
    pub PF10: PF10,
    pub PF11: PF11,
    pub PF12: PF12,
    pub PF13: PF13,
    pub PF14: PF14,
    pub PF15: PF15,
    pub PG0: PG0,
    pub PG1: PG1,
    pub PG2: PG2,
    pub PG3: PG3,
    pub PG4: PG4,
    pub PG5: PG5,
    pub PG6: PG6,
    pub PG7: PG7,
    pub PG8: PG8,
    pub PG9: PG9,
    pub PG10: PG10,
    pub PG11: PG11,
    pub PG12: PG12,
    pub PG13: PG13,
    pub PG14: PG14,
    pub PG15: PG15,
    pub PH0: PH0,
    pub PH1: PH1,
    pub PH2: PH2,
    pub PH3: PH3,
    pub PH4: PH4,
    pub PH5: PH5,
    pub PH6: PH6,
    pub PH7: PH7,
    pub PH8: PH8,
    pub PH9: PH9,
    pub PH10: PH10,
    pub PH11: PH11,
    pub PH12: PH12,
    pub PH13: PH13,
    pub PH14: PH14,
    pub PH15: PH15,
    pub PI0: PI0,
    pub PI1: PI1,
    pub PI2: PI2,
    pub PI3: PI3,
    pub PI4: PI4,
    pub PI5: PI5,
    pub PI6: PI6,
    pub PI7: PI7,
    pub PI8: PI8,
    pub PI9: PI9,
    pub PI10: PI10,
    pub PI11: PI11,
    pub PI12: PI12,
    pub PI13: PI13,
    pub PI14: PI14,
    pub PI15: PI15,
    pub PJ0: PJ0,
    pub PJ1: PJ1,
    pub PJ2: PJ2,
    pub PJ3: PJ3,
    pub PJ4: PJ4,
    pub PJ5: PJ5,
    pub PJ6: PJ6,
    pub PJ7: PJ7,
    pub PJ8: PJ8,
    pub PJ9: PJ9,
    pub PJ10: PJ10,
    pub PJ11: PJ11,
    pub PJ12: PJ12,
    pub PJ13: PJ13,
    pub PJ14: PJ14,
    pub PJ15: PJ15,
    pub PK0: PK0,
    pub PK1: PK1,
    pub PK2: PK2,
    pub PK3: PK3,
    pub PK4: PK4,
    pub PK5: PK5,
    pub PK6: PK6,
    pub PK7: PK7,
    pub PK8: PK8,
    pub PK9: PK9,
    pub PK10: PK10,
    pub PK11: PK11,
    pub PK12: PK12,
    pub PK13: PK13,
    pub PK14: PK14,
    pub PK15: PK15,
    pub I2C1: I2C1,
    pub I2C2: I2C2,
    pub I2C3: I2C3,
    pub I2C4: I2C4,
    pub IWDG1: IWDG1,
    pub JPEG: JPEG,
    pub LPUART1: LPUART1,
    pub MDIOS: MDIOS,
    pub OCTOSPI1: OCTOSPI1,
    pub OCTOSPI2: OCTOSPI2,
    pub OCTOSPIM: OCTOSPIM,
    pub PWR: PWR,
    pub MCO1: MCO1,
    pub MCO2: MCO2,
    pub RCC: RCC,
    pub RNG: RNG,
    pub RTC: RTC,
    pub SAI1: SAI1,
    pub SAI2: SAI2,
    pub SDMMC1: SDMMC1,
    pub SDMMC2: SDMMC2,
    pub SPI1: SPI1,
    pub SPI2: SPI2,
    pub SPI3: SPI3,
    pub SPI4: SPI4,
    pub SPI6: SPI6,
    pub SYSCFG: SYSCFG,
    pub TIM1: TIM1,
    pub TIM12: TIM12,
    pub TIM13: TIM13,
    pub TIM14: TIM14,
    pub TIM15: TIM15,
    pub TIM16: TIM16,
    pub TIM17: TIM17,
    pub TIM3: TIM3,
    pub TIM4: TIM4,
    pub TIM5: TIM5,
    pub TIM6: TIM6,
    pub TIM7: TIM7,
    pub TIM8: TIM8,
    pub UART4: UART4,
    pub UART5: UART5,
    pub UART7: UART7,
    pub UART8: UART8,
    pub UART9: UART9,
    pub UID: UID,
    pub USART1: USART1,
    pub USART10: USART10,
    pub USART2: USART2,
    pub USART3: USART3,
    pub USART6: USART6,
    pub USB_OTG_HS: USB_OTG_HS,
    pub WWDG1: WWDG1,
    pub EXTI0: EXTI0,
    pub EXTI1: EXTI1,
    pub EXTI2: EXTI2,
    pub EXTI3: EXTI3,
    pub EXTI4: EXTI4,
    pub EXTI5: EXTI5,
    pub EXTI6: EXTI6,
    pub EXTI7: EXTI7,
    pub EXTI8: EXTI8,
    pub EXTI9: EXTI9,
    pub EXTI10: EXTI10,
    pub EXTI11: EXTI11,
    pub EXTI12: EXTI12,
    pub EXTI13: EXTI13,
    pub EXTI14: EXTI14,
    pub EXTI15: EXTI15,
    pub DMA1_CH0: DMA1_CH0,
    pub DMA1_CH1: DMA1_CH1,
    pub DMA1_CH2: DMA1_CH2,
    pub DMA1_CH3: DMA1_CH3,
    pub DMA1_CH4: DMA1_CH4,
    pub DMA1_CH5: DMA1_CH5,
    pub DMA1_CH6: DMA1_CH6,
    pub DMA1_CH7: DMA1_CH7,
    pub DMA2_CH0: DMA2_CH0,
    pub DMA2_CH1: DMA2_CH1,
    pub DMA2_CH2: DMA2_CH2,
    pub DMA2_CH3: DMA2_CH3,
    pub DMA2_CH4: DMA2_CH4,
    pub DMA2_CH5: DMA2_CH5,
    pub DMA2_CH6: DMA2_CH6,
    pub DMA2_CH7: DMA2_CH7,
    pub BDMA1_CH0: BDMA1_CH0,
    pub BDMA1_CH1: BDMA1_CH1,
    pub BDMA1_CH2: BDMA1_CH2,
    pub BDMA1_CH3: BDMA1_CH3,
    pub BDMA1_CH4: BDMA1_CH4,
    pub BDMA1_CH5: BDMA1_CH5,
    pub BDMA1_CH6: BDMA1_CH6,
    pub BDMA1_CH7: BDMA1_CH7,
    pub BDMA2_CH0: BDMA2_CH0,
    pub BDMA2_CH1: BDMA2_CH1,
    pub BDMA2_CH2: BDMA2_CH2,
    pub BDMA2_CH3: BDMA2_CH3,
    pub BDMA2_CH4: BDMA2_CH4,
    pub BDMA2_CH5: BDMA2_CH5,
    pub BDMA2_CH6: BDMA2_CH6,
    pub BDMA2_CH7: BDMA2_CH7,
    pub PC2_C: PC2_C,
    pub PC3_C: PC3_C,

}

Struct containing all the peripheral singletons.

To obtain the peripherals, you must initialize the HAL, by calling crate::init.

Fields
ADC1: ADC1

ADC1 peripheral

ADC2: ADC2

ADC2 peripheral

ADC_COMMON: ADC_COMMON

ADC_COMMON peripheral

BDMA1: BDMA1

BDMA1 peripheral

BDMA2: BDMA2

BDMA2 peripheral

CRC: CRC

CRC peripheral

CRS: CRS

CRS peripheral

DAC1: DAC1

DAC1 peripheral

DAC2: DAC2

DAC2 peripheral

DBGMCU: DBGMCU

DBGMCU peripheral

DCMI: DCMI

DCMI peripheral

DMA1: DMA1

DMA1 peripheral

DMA2: DMA2

DMA2 peripheral

DMA2D: DMA2D

DMA2D peripheral

DMAMUX1: DMAMUX1

DMAMUX1 peripheral

DMAMUX2: DMAMUX2

DMAMUX2 peripheral

FDCAN1: FDCAN1

FDCAN1 peripheral

FDCAN2: FDCAN2

FDCAN2 peripheral

FDCANRAM: FDCANRAM

FDCANRAM peripheral

FLASH: FLASH

FLASH peripheral

FMC: FMC

FMC peripheral

PA0: PA0

PA0 peripheral

PA1: PA1

PA1 peripheral

PA2: PA2

PA2 peripheral

PA3: PA3

PA3 peripheral

PA4: PA4

PA4 peripheral

PA5: PA5

PA5 peripheral

PA6: PA6

PA6 peripheral

PA7: PA7

PA7 peripheral

PA8: PA8

PA8 peripheral

PA9: PA9

PA9 peripheral

PA10: PA10

PA10 peripheral

PA11: PA11

PA11 peripheral

PA12: PA12

PA12 peripheral

PA13: PA13

PA13 peripheral

PA14: PA14

PA14 peripheral

PA15: PA15

PA15 peripheral

PB0: PB0

PB0 peripheral

PB1: PB1

PB1 peripheral

PB2: PB2

PB2 peripheral

PB3: PB3

PB3 peripheral

PB4: PB4

PB4 peripheral

PB5: PB5

PB5 peripheral

PB6: PB6

PB6 peripheral

PB7: PB7

PB7 peripheral

PB8: PB8

PB8 peripheral

PB9: PB9

PB9 peripheral

PB10: PB10

PB10 peripheral

PB11: PB11

PB11 peripheral

PB12: PB12

PB12 peripheral

PB13: PB13

PB13 peripheral

PB14: PB14

PB14 peripheral

PB15: PB15

PB15 peripheral

PC0: PC0

PC0 peripheral

PC1: PC1

PC1 peripheral

PC2: PC2

PC2 peripheral

PC3: PC3

PC3 peripheral

PC4: PC4

PC4 peripheral

PC5: PC5

PC5 peripheral

PC6: PC6

PC6 peripheral

PC7: PC7

PC7 peripheral

PC8: PC8

PC8 peripheral

PC9: PC9

PC9 peripheral

PC10: PC10

PC10 peripheral

PC11: PC11

PC11 peripheral

PC12: PC12

PC12 peripheral

PC13: PC13

PC13 peripheral

PC14: PC14

PC14 peripheral

PC15: PC15

PC15 peripheral

PD0: PD0

PD0 peripheral

PD1: PD1

PD1 peripheral

PD2: PD2

PD2 peripheral

PD3: PD3

PD3 peripheral

PD4: PD4

PD4 peripheral

PD5: PD5

PD5 peripheral

PD6: PD6

PD6 peripheral

PD7: PD7

PD7 peripheral

PD8: PD8

PD8 peripheral

PD9: PD9

PD9 peripheral

PD10: PD10

PD10 peripheral

PD11: PD11

PD11 peripheral

PD12: PD12

PD12 peripheral

PD13: PD13

PD13 peripheral

PD14: PD14

PD14 peripheral

PD15: PD15

PD15 peripheral

PE0: PE0

PE0 peripheral

PE1: PE1

PE1 peripheral

PE2: PE2

PE2 peripheral

PE3: PE3

PE3 peripheral

PE4: PE4

PE4 peripheral

PE5: PE5

PE5 peripheral

PE6: PE6

PE6 peripheral

PE7: PE7

PE7 peripheral

PE8: PE8

PE8 peripheral

PE9: PE9

PE9 peripheral

PE10: PE10

PE10 peripheral

PE11: PE11

PE11 peripheral

PE12: PE12

PE12 peripheral

PE13: PE13

PE13 peripheral

PE14: PE14

PE14 peripheral

PE15: PE15

PE15 peripheral

PF0: PF0

PF0 peripheral

PF1: PF1

PF1 peripheral

PF2: PF2

PF2 peripheral

PF3: PF3

PF3 peripheral

PF4: PF4

PF4 peripheral

PF5: PF5

PF5 peripheral

PF6: PF6

PF6 peripheral

PF7: PF7

PF7 peripheral

PF8: PF8

PF8 peripheral

PF9: PF9

PF9 peripheral

PF10: PF10

PF10 peripheral

PF11: PF11

PF11 peripheral

PF12: PF12

PF12 peripheral

PF13: PF13

PF13 peripheral

PF14: PF14

PF14 peripheral

PF15: PF15

PF15 peripheral

PG0: PG0

PG0 peripheral

PG1: PG1

PG1 peripheral

PG2: PG2

PG2 peripheral

PG3: PG3

PG3 peripheral

PG4: PG4

PG4 peripheral

PG5: PG5

PG5 peripheral

PG6: PG6

PG6 peripheral

PG7: PG7

PG7 peripheral

PG8: PG8

PG8 peripheral

PG9: PG9

PG9 peripheral

PG10: PG10

PG10 peripheral

PG11: PG11

PG11 peripheral

PG12: PG12

PG12 peripheral

PG13: PG13

PG13 peripheral

PG14: PG14

PG14 peripheral

PG15: PG15

PG15 peripheral

PH0: PH0

PH0 peripheral

PH1: PH1

PH1 peripheral

PH2: PH2

PH2 peripheral

PH3: PH3

PH3 peripheral

PH4: PH4

PH4 peripheral

PH5: PH5

PH5 peripheral

PH6: PH6

PH6 peripheral

PH7: PH7

PH7 peripheral

PH8: PH8

PH8 peripheral

PH9: PH9

PH9 peripheral

PH10: PH10

PH10 peripheral

PH11: PH11

PH11 peripheral

PH12: PH12

PH12 peripheral

PH13: PH13

PH13 peripheral

PH14: PH14

PH14 peripheral

PH15: PH15

PH15 peripheral

PI0: PI0

PI0 peripheral

PI1: PI1

PI1 peripheral

PI2: PI2

PI2 peripheral

PI3: PI3

PI3 peripheral

PI4: PI4

PI4 peripheral

PI5: PI5

PI5 peripheral

PI6: PI6

PI6 peripheral

PI7: PI7

PI7 peripheral

PI8: PI8

PI8 peripheral

PI9: PI9

PI9 peripheral

PI10: PI10

PI10 peripheral

PI11: PI11

PI11 peripheral

PI12: PI12

PI12 peripheral

PI13: PI13

PI13 peripheral

PI14: PI14

PI14 peripheral

PI15: PI15

PI15 peripheral

PJ0: PJ0

PJ0 peripheral

PJ1: PJ1

PJ1 peripheral

PJ2: PJ2

PJ2 peripheral

PJ3: PJ3

PJ3 peripheral

PJ4: PJ4

PJ4 peripheral

PJ5: PJ5

PJ5 peripheral

PJ6: PJ6

PJ6 peripheral

PJ7: PJ7

PJ7 peripheral

PJ8: PJ8

PJ8 peripheral

PJ9: PJ9

PJ9 peripheral

PJ10: PJ10

PJ10 peripheral

PJ11: PJ11

PJ11 peripheral

PJ12: PJ12

PJ12 peripheral

PJ13: PJ13

PJ13 peripheral

PJ14: PJ14

PJ14 peripheral

PJ15: PJ15

PJ15 peripheral

PK0: PK0

PK0 peripheral

PK1: PK1

PK1 peripheral

PK2: PK2

PK2 peripheral

PK3: PK3

PK3 peripheral

PK4: PK4

PK4 peripheral

PK5: PK5

PK5 peripheral

PK6: PK6

PK6 peripheral

PK7: PK7

PK7 peripheral

PK8: PK8

PK8 peripheral

PK9: PK9

PK9 peripheral

PK10: PK10

PK10 peripheral

PK11: PK11

PK11 peripheral

PK12: PK12

PK12 peripheral

PK13: PK13

PK13 peripheral

PK14: PK14

PK14 peripheral

PK15: PK15

PK15 peripheral

I2C1: I2C1

I2C1 peripheral

I2C2: I2C2

I2C2 peripheral

I2C3: I2C3

I2C3 peripheral

I2C4: I2C4

I2C4 peripheral

IWDG1: IWDG1

IWDG1 peripheral

JPEG: JPEG

JPEG peripheral

LPUART1: LPUART1

LPUART1 peripheral

MDIOS: MDIOS

MDIOS peripheral

OCTOSPI1: OCTOSPI1

OCTOSPI1 peripheral

OCTOSPI2: OCTOSPI2

OCTOSPI2 peripheral

OCTOSPIM: OCTOSPIM

OCTOSPIM peripheral

PWR: PWR

PWR peripheral

MCO1: MCO1

MCO1 peripheral

MCO2: MCO2

MCO2 peripheral

RCC: RCC

RCC peripheral

RNG: RNG

RNG peripheral

RTC: RTC

RTC peripheral

SAI1: SAI1

SAI1 peripheral

SAI2: SAI2

SAI2 peripheral

SDMMC1: SDMMC1

SDMMC1 peripheral

SDMMC2: SDMMC2

SDMMC2 peripheral

SPI1: SPI1

SPI1 peripheral

SPI2: SPI2

SPI2 peripheral

SPI3: SPI3

SPI3 peripheral

SPI4: SPI4

SPI4 peripheral

SPI6: SPI6

SPI6 peripheral

SYSCFG: SYSCFG

SYSCFG peripheral

TIM1: TIM1

TIM1 peripheral

TIM12: TIM12

TIM12 peripheral

TIM13: TIM13

TIM13 peripheral

TIM14: TIM14

TIM14 peripheral

TIM15: TIM15

TIM15 peripheral

TIM16: TIM16

TIM16 peripheral

TIM17: TIM17

TIM17 peripheral

TIM3: TIM3

TIM3 peripheral

TIM4: TIM4

TIM4 peripheral

TIM5: TIM5

TIM5 peripheral

TIM6: TIM6

TIM6 peripheral

TIM7: TIM7

TIM7 peripheral

TIM8: TIM8

TIM8 peripheral

UART4: UART4

UART4 peripheral

UART5: UART5

UART5 peripheral

UART7: UART7

UART7 peripheral

UART8: UART8

UART8 peripheral

UART9: UART9

UART9 peripheral

UID: UID

UID peripheral

USART1: USART1

USART1 peripheral

USART10: USART10

USART10 peripheral

USART2: USART2

USART2 peripheral

USART3: USART3

USART3 peripheral

USART6: USART6

USART6 peripheral

USB_OTG_HS: USB_OTG_HS

USB_OTG_HS peripheral

WWDG1: WWDG1

WWDG1 peripheral

EXTI0: EXTI0

EXTI0 peripheral

EXTI1: EXTI1

EXTI1 peripheral

EXTI2: EXTI2

EXTI2 peripheral

EXTI3: EXTI3

EXTI3 peripheral

EXTI4: EXTI4

EXTI4 peripheral

EXTI5: EXTI5

EXTI5 peripheral

EXTI6: EXTI6

EXTI6 peripheral

EXTI7: EXTI7

EXTI7 peripheral

EXTI8: EXTI8

EXTI8 peripheral

EXTI9: EXTI9

EXTI9 peripheral

EXTI10: EXTI10

EXTI10 peripheral

EXTI11: EXTI11

EXTI11 peripheral

EXTI12: EXTI12

EXTI12 peripheral

EXTI13: EXTI13

EXTI13 peripheral

EXTI14: EXTI14

EXTI14 peripheral

EXTI15: EXTI15

EXTI15 peripheral

DMA1_CH0: DMA1_CH0

DMA1_CH0 peripheral

DMA1_CH1: DMA1_CH1

DMA1_CH1 peripheral

DMA1_CH2: DMA1_CH2

DMA1_CH2 peripheral

DMA1_CH3: DMA1_CH3

DMA1_CH3 peripheral

DMA1_CH4: DMA1_CH4

DMA1_CH4 peripheral

DMA1_CH5: DMA1_CH5

DMA1_CH5 peripheral

DMA1_CH6: DMA1_CH6

DMA1_CH6 peripheral

DMA1_CH7: DMA1_CH7

DMA1_CH7 peripheral

DMA2_CH0: DMA2_CH0

DMA2_CH0 peripheral

DMA2_CH1: DMA2_CH1

DMA2_CH1 peripheral

DMA2_CH2: DMA2_CH2

DMA2_CH2 peripheral

DMA2_CH3: DMA2_CH3

DMA2_CH3 peripheral

DMA2_CH4: DMA2_CH4

DMA2_CH4 peripheral

DMA2_CH5: DMA2_CH5

DMA2_CH5 peripheral

DMA2_CH6: DMA2_CH6

DMA2_CH6 peripheral

DMA2_CH7: DMA2_CH7

DMA2_CH7 peripheral

BDMA1_CH0: BDMA1_CH0

BDMA1_CH0 peripheral

BDMA1_CH1: BDMA1_CH1

BDMA1_CH1 peripheral

BDMA1_CH2: BDMA1_CH2

BDMA1_CH2 peripheral

BDMA1_CH3: BDMA1_CH3

BDMA1_CH3 peripheral

BDMA1_CH4: BDMA1_CH4

BDMA1_CH4 peripheral

BDMA1_CH5: BDMA1_CH5

BDMA1_CH5 peripheral

BDMA1_CH6: BDMA1_CH6

BDMA1_CH6 peripheral

BDMA1_CH7: BDMA1_CH7

BDMA1_CH7 peripheral

BDMA2_CH0: BDMA2_CH0

BDMA2_CH0 peripheral

BDMA2_CH1: BDMA2_CH1

BDMA2_CH1 peripheral

BDMA2_CH2: BDMA2_CH2

BDMA2_CH2 peripheral

BDMA2_CH3: BDMA2_CH3

BDMA2_CH3 peripheral

BDMA2_CH4: BDMA2_CH4

BDMA2_CH4 peripheral

BDMA2_CH5: BDMA2_CH5

BDMA2_CH5 peripheral

BDMA2_CH6: BDMA2_CH6

BDMA2_CH6 peripheral

BDMA2_CH7: BDMA2_CH7

BDMA2_CH7 peripheral

PC2_C: PC2_C

PC2_C peripheral

PC3_C: PC3_C

PC3_C peripheral

Implementations

impl Peripherals

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.