Struct embassy_stm32::timer::complementary_pwm::ComplementaryPwm

pub struct ComplementaryPwm<'d, T> { /* private fields */ }

PWM driver with support for standard and complementary outputs.

Implementations

impl<'d, T: ComplementaryCaptureCompare16bitInstance> ComplementaryPwm<'d, T>

pub fn new( tim: impl Peripheral<P = T> + 'd, _ch1: Option<PwmPin<'d, T, Ch1>>, _ch1n: Option<ComplementaryPwmPin<'d, T, Ch1>>, _ch2: Option<PwmPin<'d, T, Ch2>>, _ch2n: Option<ComplementaryPwmPin<'d, T, Ch2>>, _ch3: Option<PwmPin<'d, T, Ch3>>, _ch3n: Option<ComplementaryPwmPin<'d, T, Ch3>>, _ch4: Option<PwmPin<'d, T, Ch4>>, _ch4n: Option<ComplementaryPwmPin<'d, T, Ch4>>, freq: Hertz, counting_mode: CountingMode ) -> Self

Create a new complementary PWM driver.

pub fn enable(&mut self, channel: Channel)

Enable the given channel.

pub fn disable(&mut self, channel: Channel)

Disable the given channel.

pub fn set_frequency(&mut self, freq: Hertz)

Set PWM frequency.

Note: when you call this, the max duty value changes, so you will have to call set_duty on all channels with the duty calculated based on the new max duty.

pub fn get_max_duty(&self) -> u16

Get max duty value.

This value depends on the configured frequency and the timer’s clock rate from RCC.

pub fn set_duty(&mut self, channel: Channel, duty: u16)

Set the duty for a given channel.

The value ranges from 0 for 0% duty, to get_max_duty for 100% duty, both included.

pub fn set_polarity(&mut self, channel: Channel, polarity: OutputPolarity)

Set the output polarity for a given channel.

pub fn set_dead_time(&mut self, value: u16)

Set the dead time as a proportion of max_duty