Struct embassy_stm32::usart::UartRx

pub struct UartRx<'d, T: BasicInstance, RxDma = NoDma> { /* private fields */ }

Rx-only UART Driver

Implementations

impl<'d, T: BasicInstance, RxDma: RxDma<T>> UartRx<'d, T, RxDma>

pub fn into_ring_buffered( self, dma_buf: &'d mut [u8] ) -> RingBufferedUartRx<'d, T, RxDma>

Turn the UartRx into a buffered uart which can continously receive in the background without the possibility of losing bytes. The dma_buf is a buffer registered to the DMA controller, and must be large enough to prevent overflows.

impl<'d, T: BasicInstance, RxDma> UartRx<'d, T, RxDma>

pub fn new( peri: impl Peripheral<P = T> + 'd, _irq: impl Binding<T::Interrupt, InterruptHandler<T>> + 'd, rx: impl Peripheral<P = impl RxPin<T>> + 'd, rx_dma: impl Peripheral<P = RxDma> + 'd, config: Config ) -> Result<Self, ConfigError>

Useful if you only want Uart Rx. It saves 1 pin and consumes a little less power.

pub fn new_with_rts( peri: impl Peripheral<P = T> + 'd, _irq: impl Binding<T::Interrupt, InterruptHandler<T>> + 'd, rx: impl Peripheral<P = impl RxPin<T>> + 'd, rts: impl Peripheral<P = impl RtsPin<T>> + 'd, rx_dma: impl Peripheral<P = RxDma> + 'd, config: Config ) -> Result<Self, ConfigError>

Create a new rx-only UART with a request-to-send pin

pub fn set_config(&mut self, config: &Config) -> Result<(), ConfigError>

Reconfigure the driver

pub async fn read(&mut self, buffer: &mut [u8]) -> Result<(), Error>
where RxDma: RxDma<T>,

Initiate an asynchronous UART read

pub fn nb_read(&mut self) -> Result<u8, Error<Error>>

Read a single u8 if there is one available, otherwise return WouldBlock

pub fn blocking_read(&mut self, buffer: &mut [u8]) -> Result<(), Error>

Perform a blocking read into buffer

pub async fn read_until_idle( &mut self, buffer: &mut [u8] ) -> Result<usize, Error>
where RxDma: RxDma<T>,

Initiate an asynchronous read with idle line detection enabled