Struct embassy_stm32::usb_otg::Driver

pub struct Driver<'d, T: Instance> { /* private fields */ }

USB driver.

Implementations

impl<'d, T: Instance> Driver<'d, T>

pub fn new_fs( _peri: impl Peripheral<P = T> + 'd, _irq: impl Binding<T::Interrupt, InterruptHandler<T>> + 'd, dp: impl Peripheral<P = impl DpPin<T>> + 'd, dm: impl Peripheral<P = impl DmPin<T>> + 'd, ep_out_buffer: &'d mut [u8], config: Config ) -> Self

Initializes USB OTG peripheral with internal Full-Speed PHY.

Arguments

ep_out_buffer - An internal buffer used to temporarily store recevied packets. Must be large enough to fit all OUT endpoint max packet sizes. Endpoint allocation will fail if it is too small.

pub fn new_hs_ulpi( _peri: impl Peripheral<P = T> + 'd, _irq: impl Binding<T::Interrupt, InterruptHandler<T>> + 'd, ulpi_clk: impl Peripheral<P = impl UlpiClkPin<T>> + 'd, ulpi_dir: impl Peripheral<P = impl UlpiDirPin<T>> + 'd, ulpi_nxt: impl Peripheral<P = impl UlpiNxtPin<T>> + 'd, ulpi_stp: impl Peripheral<P = impl UlpiStpPin<T>> + 'd, ulpi_d0: impl Peripheral<P = impl UlpiD0Pin<T>> + 'd, ulpi_d1: impl Peripheral<P = impl UlpiD1Pin<T>> + 'd, ulpi_d2: impl Peripheral<P = impl UlpiD2Pin<T>> + 'd, ulpi_d3: impl Peripheral<P = impl UlpiD3Pin<T>> + 'd, ulpi_d4: impl Peripheral<P = impl UlpiD4Pin<T>> + 'd, ulpi_d5: impl Peripheral<P = impl UlpiD5Pin<T>> + 'd, ulpi_d6: impl Peripheral<P = impl UlpiD6Pin<T>> + 'd, ulpi_d7: impl Peripheral<P = impl UlpiD7Pin<T>> + 'd, ep_out_buffer: &'d mut [u8], config: Config ) -> Self

Initializes USB OTG peripheral with external High-Speed PHY.

Arguments

ep_out_buffer - An internal buffer used to temporarily store recevied packets. Must be large enough to fit all OUT endpoint max packet sizes. Endpoint allocation will fail if it is too small.